requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
PyJWT==2.1.0
requests==2.25.1
tornado==6.1
urllib3==1.26.5
websocket-client==1.1.0
